refactor: common-service
