+++ packages/common-service/CHANGELOG.md
@@ -1,0 +1,6 @@
+# CHANGELOG
+
+## 0.2.0
+
+- refactor: clearer methods division
+- refactor: move methods which are about user to @iceworks/user-service
