+++ packages/common-service/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@iceworks/common-service",
-  "version": "0.1.31",
+  "version": "0.2.0",
   "description": "Iceworks common service for VSCode extension.",
   "files": [
     "lib"
