+++ packages/common-service/src/constants.ts
@@ -1,0 +1,12 @@
+export const CONFIGURATION_SECTION = 'iceworks';
+export const CONFIGURATION_KEY_PACKAGE_MANAGER = 'packageManager';
+export const CONFIGURATION_KEY_NPM_REGISTRY = 'npmRegistry';
+export const CONFIGURATION_KEY_MATERIAL_SOURCES = 'materialSources';
+export const CONFIGURATION_KEY_GENERATE_PAGE_PATH = 'generatePagePath';
+export const CONFIGURATION_KEY_GENERATE_COMPONENT_PATH = 'generateComponentPath';
+export const CONFIGURATION_SECTION_PACKAGE_MANAGER = `${CONFIGURATION_SECTION}.${CONFIGURATION_KEY_PACKAGE_MANAGER}`;
+export const CONFIGURATION_SECTION_NPM_REGISTRY = `${CONFIGURATION_SECTION}.${CONFIGURATION_KEY_NPM_REGISTRY}`;
+export const CONFIGURATION_SECTION_MATERIAL_SOURCES = `${CONFIGURATION_SECTION}.${CONFIGURATION_KEY_MATERIAL_SOURCES}`;
+
+export const indexFileSuffix = ['.jsx', '.js', '.tsx', '.ts', '.rml', '.vue'];
+export const didShowWelcomePageStateKey = 'iceworks.didShowWelcomePage';
