+++ packages/user-service/CHANGELOG.md
@@ -1,0 +1,5 @@
+# CHANGELOG
+
+## 0.1.0
+
+- feat: init
